pyproject.toml
[project]
name = "nanobanana"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.5",
    "keyboard>=0.13.5",
    "pillow-heif>=1.1.0",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
]
